ModuleHandler god class divided into logical units
